requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cytoolz==0.10.1
distlib==0.3.0
dydx-python==0.8.0
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.1
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
filelock==3.0.12
google-auth==1.11.3
google-auth-httplib2==0.0.3
gspread==3.3.0
hexbytes==0.2.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
ipfshttpclient==0.4.12
jsonschema==3.2.0
lru-dict==1.1.6
more-itertools==8.2.0
multiaddr==0.0.9
netaddr==0.7.19
numpy==1.18.2
oauth2client==4.1.3
packaging==20.3
pandas==1.0.3
parsimonious==0.8.1
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pyparsing==2.4.6
pypiwin32==223
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
requests==2.23.0
requests-mock==1.7.0
rlp==1.2.0
rsa==4.0
six==1.14.0
toml==0.10.0
toolz==0.10.0
tox==3.14.5
typing-extensions==3.7.4.1
urllib3==1.25.8
varint==1.0.2
virtualenv==20.0.13
wcwidth==0.1.8
web3==5.7.0
websockets==8.1
zipp==3.1.0
